requirements.txt
fastapi ==  0.65.3
uvicorn == 0.14.0
asyncpg == 0.23.0
tortoise-orm == 0.17.4
aerich == 0.5.3
pytest == 6.2.4
requests == 2.25.1